Set event_type to cast for ra_log_event
The ra_node_proc crashes due to a bad action event type given to
gen_statem. This appears to be caused when we omit the
event_type from the effects returns from ra_node:handle_leader/2 which
is later assumed to be
be a call action. This fails when the ra_node_proc is in a `leader`
state and receives a DOWN info message from the customer monitor. The
ra_effects are assumed to be call actions when in fact it should be a
cast. Here, we explicitly set the event type to cast.